Refactoring operation name : extract class
Refactoring object : MakeBytesToSegment

Reason : makeNumeric, changeNumericToSegment, containNonNumericCharaters
has independent role.
So, make these function to class(MakeNumericToSegment).
